requirements.txt
accelerate
diffusers
huggingface-hub
python-dotenv
torch
transformers
transformers[sentencepiece]